Fixed converison from wxString to char* which in cases returned garbled text. Use size_t instead of int

+++ tools/source_maker/source_templates/functions.cpp
@@ -76,7 +76,7 @@ zval * wxphp_sprintf(INTERNAL_FUNCTION_PARAMETERS, int offset = 0)
 PHP_FUNCTION(php_wxExecute)
 {
     char* _argStr0;
-    int _argStr0_len;
+    size_t _argStr0_len;
 
     long flags = 0;
     zval z_process;
